Make two layouts: one main and one for docs. Use these layouts in  Gatsby's API to not make full rerender of page on navigation. It fixes transition in doc's sidebar
