+++ gatsby-browser.js
@@ -1,0 +1,5 @@
+/* eslint-env node */
+
+const PageWrapper = require('./src/components/PageWrapper').default
+
+exports.wrapPageElement = PageWrapper
