+++ src/components/Documentation/styles.js
@@ -1,132 +1,5 @@
 import styled from 'styled-components'
 
-import { media } from '../../styles'
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: row;
-  max-width: 1200px;
-  margin: 0 auto;
-  background: white;
-  z-index: 2;
-
-  &:before {
-    content: '';
-    display: block;
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    width: 50%;
-    background-color: #eef4f8;
-    z-index: -1;
-    pointer-events: none;
-  }
-`
-
-export const Backdrop = styled.div`
-  display: none;
-
-  ${media.phablet`
-    display: block;
-    opacity: 0;
-    pointer-events: none;
-    transition: opacity .3s linear;
-
-    ${props =>
-      props.visible &&
-      `
-      content: '';
-      position: fixed;
-      top: 0;
-      left: 0;
-      bottom: 0;
-      right: 0;
-      background-color: rgba(0, 0, 0, 0.4);
-      z-index: 1;
-      opacity: 1;
-      pointer-events: all;
-    `}
-  `};
-`
-
-export const Side = styled.div`
-  width: 280px;
-  background-color: #eef4f8;
-
-  @media only screen and (max-width: 1200px) {
-    padding-left: 15px;
-  }
-
-  ${media.phablet`
-    position: fixed;
-    display: block;
-    z-index: 2;
-    top: 78px;
-    bottom: 0;
-    left: 0;
-    right: 60px;
-    box-shadow: rgba(0, 0, 0, 0.14) 0px 0px 4px, rgba(0, 0, 0, 0.28) 0px 4px 8px;
-    transform: translateX(-110%);
-    transition: transform .35s ease;
-
-    ${props =>
-      props.isOpen &&
-      `
-      transform: translateX(0);
-    `}
-  `};
-`
-
-export const SearchArea = styled.div`
-  height: 60px;
-  display: flex;
-  align-items: center;
-  background-color: #eef4f8;
-  z-index: 10;
-  position: sticky;
-  top: 0;
-
-  ${media.phablet`
-    position: relative;
-    padding: 0 20px;
-  `};
-
-  form {
-    height: 40px;
-  }
-`
-
-export const SideToggle = styled.div`
-  display: none;
-  position: fixed;
-  z-index: 2;
-  left: 8px;
-  bottom: 20px;
-  width: 45px;
-  height: 45px;
-  border-radius: 50%;
-  background-color: rgba(255, 255, 255, 0.9);
-  box-shadow: 0 0px 9px 0 rgba(0, 0, 0, 0.15);
-  transition: transform 0.3s ease;
-  justify-content: center;
-  align-items: center;
-
-  ${media.phablet`
-    display: flex;
-
-    > div {
-      transform: scale(0.75);
-    }
-  `};
-
-  ${({ isMenuOpen }) =>
-    isMenuOpen &&
-    `
-    transform: translateX(calc(100vw - 60px));
-  `};
-`
-
 export const LightButton = styled.a`
   display: inline-flex;
   justify-content: center;
