+++ src/components/Home/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import LearnMore from '../LearnMore'
-import Page from '../Page'
 import Hero from '../Hero'
 import LandingHero from '../LandingHero'
 import Diagram from '../Diagram'
@@ -13,7 +12,7 @@ import { LearnMoreSection } from './styles'
 
 export default function HomePage() {
   return (
-    <Page stickHeader={true}>
+    <>
       <Hero>
         <LandingHero />
         <LearnMoreSection>
@@ -25,6 +24,6 @@ export default function HomePage() {
       <PromoSection />
       <UseCases />
       <Subscribe />
-    </Page>
+    </>
   )
 }
