+++ src/components/Layout/index.js
@@ -1,30 +1,50 @@
-import React from 'react'
+import React, { useEffect } from 'react'
+import { useLocation } from '@reach/router'
 import PropTypes from 'prop-types'
+import { GlobalStyle } from '../../styles'
 
-import TopMenu from '../TopMenu'
-import Footer from '../Footer'
-import HamburgerMenu from '../HamburgerMenu'
-
-import { Wrapper, Bodybag, ModalRoot } from './styles'
+import MainLayout from '../MainLayout'
+import DocLayout from '../DocLayout'
 
 import './fonts/fonts.css'
 
-export default function Layout({ children, enableSmoothScroll, isDocPage }) {
+const useAnchorNavigation = () => {
+  const location = useLocation()
+
+  useEffect(() => {
+    if (location.hash) {
+      const node = document.querySelector(location.hash)
+
+      if (node) {
+        node.scrollIntoView()
+      }
+    } else {
+      document
+        .getElementById('bodybag')
+        .scrollTo({ top: 0, behavior: 'smooth' })
+    }
+  }, [location.href])
+}
+
+export default function Layout(props) {
+  let LayoutComponent = MainLayout
+
+  useAnchorNavigation()
+
+  if (/^\/doc/.test(props.location.pathname)) {
+    LayoutComponent = DocLayout
+  }
+
   return (
-    <Wrapper>
-      <TopMenu isDocPage={isDocPage} />
-      <HamburgerMenu />
-      <Bodybag id="bodybag" enableSmoothScroll={enableSmoothScroll}>
-        {children}
-        <Footer isDocPage={isDocPage} />
-      </Bodybag>
-      <ModalRoot id="modal-root"></ModalRoot>
-    </Wrapper>
+    <>
+      <GlobalStyle />
+      <LayoutComponent {...props} />
+    </>
   )
 }
 
 Layout.propTypes = {
-  children: PropTypes.node.isRequired,
-  enableSmoothScroll: PropTypes.bool,
-  isDocPage: PropTypes.bool
+  location: PropTypes.shape({
+    pathname: PropTypes.string.isRequired
+  })
 }
